update flatbuffers to 1.12.0

+++ WildRPCClient/WildRPCClient/include/Parameters.h
@@ -11,7 +11,7 @@
 
 namespace WildRPC
 {
-    enum Type : int;
+    enum Type : int8_t;
 }
 
 namespace WRPC
